Fix: Split oversized WarpGraph specs
