+++ test/unit/domain/WarpGraph.strands.test.ts
@@ -791,649 +791,4 @@ describe('WarpCore strand foundation', () => {
     });
   });
 
-  it('compareStrand reports strand-vs-base divergence as substrate facts', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'n1', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'n1', key: 'status', value: 'base' },
-      ],
-      reads: [],
-      writes: ['n1'],
-    });
-
-    await graph.createStrand({
-      strandId: 'ws_compare_base',
-      owner: 'alice',
-    });
-
-    const overlaySha = await graph.patchStrand('ws_compare_base', (p) => {
-      p.setProperty('n1', 'status', 'overlay');
-    });
-
-    const comparison = await graph.compareStrand('ws_compare_base', { targetId: 'n1' });
-
-    expect(comparison.comparisonVersion).toBe('coordinate-compare/v1');
-    expect(typeof comparison.comparisonDigest).toBe('string');
-    expect(comparison.left.requested).toEqual({
-      kind: 'strand',
-      strandId: 'ws_compare_base',
-    });
-    expect(comparison.right.requested).toMatchObject({
-      kind: 'strand_base',
-      strandId: 'ws_compare_base',
-    });
-    expect(comparison.visiblePatchDivergence).toEqual({
-      sharedCount: 1,
-      leftOnlyCount: 1,
-      rightOnlyCount: 0,
-      leftOnlyPatchShas: [overlaySha],
-      rightOnlyPatchShas: [],
-      target: {
-        targetId: 'n1',
-        leftCount: 2,
-        rightCount: 1,
-        sharedCount: 1,
-        leftOnlyCount: 1,
-        rightOnlyCount: 0,
-        leftOnlyPatchShas: [overlaySha],
-        rightOnlyPatchShas: [],
-      },
-    });
-    expect(comparison.visibleState.changed).toBe(true);
-    expect(comparison.visibleState.nodeProperties.changed).toEqual([
-      { node: 'n1', key: 'status', leftValue: 'overlay', rightValue: 'base' },
-    ]);
-    expect(comparison.visibleState.target).toMatchObject({
-      targetId: 'n1',
-      changed: true,
-      propertyDelta: {
-        added: [],
-        removed: [],
-        changed: [{ key: 'status', leftValue: 'overlay', rightValue: 'base' }],
-      },
-    });
-  });
-
-  it('compareStrand supports live-frontier comparisons without mutating the strand boundary', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'n1', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'n1', key: 'status', value: 'base' },
-      ],
-      reads: [],
-      writes: ['n1'],
-    });
-
-    await graph.createStrand({
-      strandId: 'ws_compare_live',
-      owner: 'alice',
-    });
-
-    const overlaySha = await graph.patchStrand('ws_compare_live', (p) => {
-      p.setProperty('n1', 'status', 'overlay');
-    });
-    const liveSha = await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 2,
-      context: new Map([['alice', 1]]),
-      ops: [
-        { type: 'PropSet', node: 'n1', key: 'status', value: 'live' },
-      ],
-      reads: ['n1'],
-      writes: ['n1'],
-    });
-
-    const comparison = await graph.compareStrand('ws_compare_live', {
-      against: 'live',
-      targetId: 'n1',
-    });
-
-    expect(comparison.right.requested).toEqual({
-      kind: 'live',
-    });
-    expect(comparison.visiblePatchDivergence.leftOnlyPatchShas).toEqual([overlaySha]);
-    expect(comparison.visiblePatchDivergence.rightOnlyPatchShas).toEqual([liveSha]);
-    expect(comparison.visiblePatchDivergence.sharedCount).toBe(1);
-    expect(comparison.visibleState.nodeProperties.changed).toEqual([
-      { node: 'n1', key: 'status', leftValue: 'overlay', rightValue: 'live' },
-    ]);
-  });
-
-  it('compareStrand supports strand-vs-strand comparisons and compareCoordinates handles explicit coordinates', async () => {
-    const redSha = await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'n1', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'n1', key: 'color', value: 'red' },
-      ],
-      reads: [],
-      writes: ['n1'],
-    });
-
-    await graph.createStrand({
-      strandId: 'ws_left',
-      owner: 'alice',
-    });
-    await graph.createStrand({
-      strandId: 'ws_right',
-      owner: 'alice',
-    });
-
-    const leftOverlaySha = await graph.patchStrand('ws_left', (p) => {
-      p.setProperty('n1', 'color', 'blue');
-    });
-    const rightOverlaySha = await graph.patchStrand('ws_right', (p) => {
-      p.setProperty('n1', 'color', 'green');
-    });
-
-    const wsComparison = await graph.compareStrand('ws_left', {
-      against: { kind: 'strand', strandId: 'ws_right' },
-      targetId: 'n1',
-    });
-
-    expect(wsComparison.visiblePatchDivergence).toEqual({
-      sharedCount: 1,
-      leftOnlyCount: 1,
-      rightOnlyCount: 1,
-      leftOnlyPatchShas: [leftOverlaySha],
-      rightOnlyPatchShas: [rightOverlaySha],
-      target: {
-        targetId: 'n1',
-        leftCount: 2,
-        rightCount: 2,
-        sharedCount: 1,
-        leftOnlyCount: 1,
-        rightOnlyCount: 1,
-        leftOnlyPatchShas: [leftOverlaySha],
-        rightOnlyPatchShas: [rightOverlaySha],
-      },
-    });
-    expect(wsComparison.visibleState.nodeProperties.changed).toEqual([
-      { node: 'n1', key: 'color', leftValue: 'blue', rightValue: 'green' },
-    ]);
-
-    const frontierAtRed = { alice: redSha };
-    const blueSha = await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 2,
-      context: new Map([['alice', 1]]),
-      ops: [
-        { type: 'PropSet', node: 'n1', key: 'color', value: 'blue' },
-      ],
-      reads: ['n1'],
-      writes: ['n1'],
-    });
-
-    const coordinateComparison = await graph.compareCoordinates({
-      left: { kind: 'coordinate', frontier: frontierAtRed },
-      right: { kind: 'live' },
-      targetId: 'n1',
-    });
-
-    expect(coordinateComparison.left.requested).toEqual({
-      kind: 'coordinate',
-      frontier: frontierAtRed,
-      ceiling: null,
-    });
-    expect(coordinateComparison.visiblePatchDivergence.leftOnlyPatchShas).toEqual([]);
-    expect(coordinateComparison.visiblePatchDivergence.rightOnlyPatchShas).toEqual([blueSha]);
-    expect(coordinateComparison.visibleState.nodeProperties.changed).toEqual([
-      { node: 'n1', key: 'color', leftValue: 'red', rightValue: 'blue' },
-    ]);
-
-    const factExport = exportCoordinateComparisonFact(coordinateComparison);
-    expect(factExport).toEqual({
-      exportVersion: 'coordinate-comparison-fact/v1',
-      factKind: 'coordinate-comparison',
-      factDigest: coordinateComparison.comparisonDigest,
-      canonicalFactJson: expect.any(String),
-      fact: {
-        comparisonVersion: coordinateComparison.comparisonVersion,
-        left: coordinateComparison.left,
-        right: coordinateComparison.right,
-        visiblePatchDivergence: coordinateComparison.visiblePatchDivergence,
-        visibleState: coordinateComparison.visibleState,
-      },
-    });
-    expect(JSON.parse(factExport.canonicalFactJson)).toEqual(factExport.fact);
-    await expect(graph._crypto.hash('sha256', factExport.canonicalFactJson)).resolves.toBe(factExport.factDigest);
-  });
-
-  it('scopes coordinate comparison and transfer planning by node-id prefix without mutating the raw substrate truth', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'task:1', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'task:1', key: 'status', value: 'ready' },
-      ],
-      reads: [],
-      writes: ['task:1'],
-    });
-
-    const operationalFrontier = Object.fromEntries(await graph.getFrontier());
-
-    const governanceSha = await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 2,
-      context: new Map([['alice', 1]]),
-      ops: [
-        { type: 'NodeAdd', node: 'comparison-artifact:cmp-1', dot: Dot.create('alice', 2) },
-        { type: 'PropSet', node: 'comparison-artifact:cmp-1', key: 'kind', value: 'comparison-artifact' },
-        { type: 'EdgeAdd', from: 'task:1', to: 'comparison-artifact:cmp-1', label: 'governs', dot: Dot.create('alice', 3) },
-      ],
-      reads: ['task:1', 'comparison-artifact:cmp-1'],
-      writes: ['comparison-artifact:cmp-1', 'task:1\0comparison-artifact:cmp-1\0governs'],
-    });
-
-    const rawComparison = await graph.compareCoordinates({
-      left: { kind: 'coordinate', frontier: operationalFrontier },
-      right: { kind: 'live' },
-    });
-    expect(rawComparison.visiblePatchDivergence.rightOnlyPatchShas).toEqual([governanceSha]);
-    expect(rawComparison.visibleState.changed).toBe(true);
-
-    const scope = {
-      nodeIdPrefixes: {
-        include: [],
-        exclude: ['comparison-artifact:'],
-      },
-    };
-
-    const scopedComparison = await graph.compareCoordinates({
-      left: { kind: 'coordinate', frontier: operationalFrontier },
-      right: { kind: 'live' },
-      scope,
-    });
-
-    expect(scopedComparison.scope).toEqual(scope);
-    expect(scopedComparison.visibleState.changed).toBe(false);
-    expect(scopedComparison.visiblePatchDivergence.leftOnlyCount).toBe(0);
-    expect(scopedComparison.visiblePatchDivergence.rightOnlyCount).toBe(0);
-    expect(scopedComparison.left.resolved.patchUniverseDigest).toBe(scopedComparison.right.resolved.patchUniverseDigest);
-
-    const scopedFactExport = exportCoordinateComparisonFact(scopedComparison);
-    expect(scopedFactExport.fact.scope).toEqual(scope);
-    expect(JSON.parse(scopedFactExport.canonicalFactJson)).toEqual(scopedFactExport.fact);
-    await expect(graph._crypto.hash('sha256', scopedFactExport.canonicalFactJson)).resolves.toBe(scopedFactExport.factDigest);
-
-    const scopedTransfer = await graph.planCoordinateTransfer({
-      source: { kind: 'coordinate', frontier: operationalFrontier },
-      target: { kind: 'live' },
-      scope,
-    });
-
-    expect(scopedTransfer.scope).toEqual(scope);
-    expect(scopedTransfer.changed).toBe(false);
-    expect(scopedTransfer.summary.opCount).toBe(0);
-    expect(scopedTransfer.ops).toEqual([]);
-
-    const transferFactExport = exportCoordinateTransferPlanFact(scopedTransfer);
-    expect(transferFactExport.fact.scope).toEqual(scope);
-  });
-
-  it('planStrandTransfer emits a deterministic transfer plan including property clears and content attachment updates', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'doc:1', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'doc:1', key: 'status', value: 'draft' },
-        { type: 'PropSet', node: 'doc:1', key: 'obsolete', value: true },
-      ],
-      reads: [],
-      writes: ['doc:1'],
-    });
-
-    await graph.materialize();
-    await graph.patch(async (p) => {
-      await p.attachContent('doc:1', 'live-body', { mime: 'text/plain', size: 9 });
-    });
-
-    await graph.createStrand({
-      strandId: 'ws_transfer_live',
-      owner: 'alice',
-    });
-
-    await graph.patchStrand('ws_transfer_live', async (p) => {
-      p.setProperty('doc:1', 'status', 'ready');
-      p.setProperty('doc:1', 'obsolete', null);
-      await p.attachContent('doc:1', 'worldline-body', { mime: 'text/plain', size: 14 });
-    });
-
-    const strandState = await graph.materializeStrand('ws_transfer_live');
-    const strandReader = createStateReader(strandState);
-    const strandContentMeta = strandReader.getNodeContentMeta('doc:1');
-    expect(strandContentMeta).toEqual({
-      oid: expect.any(String),
-      mime: 'text/plain',
-      size: 14,
-    });
-    persistence._blobs.set(
-      /** @type {{ oid: string }} */ (strandContentMeta).oid,
-      Buffer.from('worldline-body'),
-    );
-
-    const transferPlan = await graph.planStrandTransfer('ws_transfer_live');
-
-    expect(transferPlan.transferVersion).toBe('coordinate-transfer-plan/v1');
-    expect(typeof transferPlan.transferDigest).toBe('string');
-    expect(transferPlan.comparisonDigest).toEqual(expect.any(String));
-    expect(transferPlan.source.requested).toEqual({
-      kind: 'strand',
-      strandId: 'ws_transfer_live',
-    });
-    expect(transferPlan.target.requested).toEqual({
-      kind: 'live',
-    });
-    expect(transferPlan.summary).toMatchObject({
-      opCount: 3,
-      setNodePropertyCount: 1,
-      clearNodePropertyCount: 1,
-      attachNodeContentCount: 1,
-      clearNodeContentCount: 0,
-    });
-    expect(transferPlan.ops).toContainEqual({
-      op: 'set_node_property',
-      nodeId: 'doc:1',
-      key: 'status',
-      value: 'ready',
-    });
-    expect(transferPlan.ops).toContainEqual({
-      op: 'set_node_property',
-      nodeId: 'doc:1',
-      key: 'obsolete',
-      value: null,
-    });
-    const attachOp = /** @type {import('../../../index.ts').VisibleStateTransferOperationV1 & { op: 'attach_node_content', nodeId: string, content: Uint8Array, contentOid: string, mime?: string|null, size?: number|null }} */ (
-      transferPlan.ops.find((op) => op.op === 'attach_node_content' && op.nodeId === 'doc:1')
-    );
-    expect(attachOp).toMatchObject({
-      op: 'attach_node_content',
-      nodeId: 'doc:1',
-      contentOid: expect.any(String),
-      mime: 'text/plain',
-      size: 14,
-    });
-    expect(Buffer.from(attachOp.content).toString('utf8')).toBe('worldline-body');
-
-    const factExport = exportCoordinateTransferPlanFact(transferPlan);
-    expect(factExport).toEqual({
-      exportVersion: 'coordinate-transfer-plan-fact/v1',
-      factKind: 'coordinate-transfer-plan',
-      factDigest: transferPlan.transferDigest,
-      canonicalFactJson: expect.any(String),
-      fact: expect.objectContaining({
-        transferVersion: transferPlan.transferVersion,
-        comparisonDigest: transferPlan.comparisonDigest,
-        changed: transferPlan.changed,
-        source: transferPlan.source,
-        target: transferPlan.target,
-        summary: transferPlan.summary,
-        ops: expect.arrayContaining([
-          expect.objectContaining({
-            op: 'set_node_property',
-            nodeId: 'doc:1',
-            key: 'status',
-            value: 'ready',
-          }),
-          expect.objectContaining({
-            op: 'set_node_property',
-            nodeId: 'doc:1',
-            key: 'obsolete',
-            value: null,
-          }),
-          {
-            op: 'attach_node_content',
-            nodeId: 'doc:1',
-            contentOid: attachOp.contentOid,
-            mime: 'text/plain',
-            size: 14,
-          },
-        ]),
-      }),
-    });
-    expect(JSON.parse(factExport.canonicalFactJson)).toEqual(factExport.fact);
-    await expect(graph._crypto.hash('sha256', factExport.canonicalFactJson)).resolves.toBe(factExport.factDigest);
-  });
-
-  it('planStrandTransfer includes braided support visibility in the candidate transfer plan', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'task:1', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'task:1', key: 'status', value: 'base' },
-      ],
-      reads: [],
-      writes: ['task:1'],
-    });
-
-    await graph.createStrand({ strandId: 'ws_target', owner: 'alice' });
-    await graph.createStrand({ strandId: 'ws_support', owner: 'alice' });
-
-    await graph.patchStrand('ws_target', (p) => {
-      p.setProperty('task:1', 'status', 'target');
-    });
-    await graph.patchStrand('ws_support', (p) => {
-      p.addNode('task:2');
-      p.setProperty('task:2', 'kind', 'support');
-    });
-
-    await graph.braidStrand('ws_target', {
-      braidedStrandIds: ['ws_support'],
-      writable: false,
-    });
-
-    const transferPlan = await graph.planStrandTransfer('ws_target');
-
-    expect(transferPlan.source.resolved.strand?.braid.braidedStrandIds).toEqual(['ws_support']);
-    expect(transferPlan.ops).toContainEqual({
-      op: 'set_node_property',
-      nodeId: 'task:1',
-      key: 'status',
-      value: 'target',
-    });
-    expect(transferPlan.ops).toContainEqual({
-      op: 'add_node',
-      nodeId: 'task:2',
-    });
-    expect(transferPlan.ops).toContainEqual({
-      op: 'set_node_property',
-      nodeId: 'task:2',
-      key: 'kind',
-      value: 'support',
-    });
-  });
-
-  it('planStrandTransfer represents content removals as explicit clear operations', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'doc:clear', dot: Dot.create('alice', 1) },
-      ],
-      reads: [],
-      writes: ['doc:clear'],
-    });
-
-    await graph.materialize();
-    await graph.patch(async (p) => {
-      await p.attachContent('doc:clear', 'live-only', { mime: 'text/plain', size: 9 });
-    });
-
-    await graph.createStrand({
-      strandId: 'ws_clear_content',
-      owner: 'alice',
-    });
-
-    await graph.patchStrand('ws_clear_content', (p) => {
-      p.clearContent('doc:clear');
-    });
-
-    const transferPlan = await graph.planStrandTransfer('ws_clear_content');
-
-    expect(transferPlan.summary).toMatchObject({
-      opCount: 1,
-      clearNodeContentCount: 1,
-    });
-    expect(transferPlan.ops).toEqual([
-      {
-        op: 'clear_node_content',
-        nodeId: 'doc:clear',
-      },
-    ]);
-  });
-
-  it('queues strand intents without mutating visible state or live truth', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'task:queued', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'task:queued', key: 'status', value: 'base' },
-      ],
-      reads: [],
-      writes: ['task:queued'],
-    });
-
-    await graph.createStrand({ strandId: 'ws_queue', owner: 'alice' });
-
-    const queued = await graph.queueStrandIntent('ws_queue', (p) => {
-      p.setProperty('task:queued', 'status', 'queued');
-    });
-
-    expect(queued.intentId).toBe('ws_queue.intent.0001');
-    expect(queued.patch.writes).toEqual(['task:queued']);
-
-    const intents = await graph.listStrandIntents('ws_queue');
-    expect(intents.map((intent) => intent.intentId)).toEqual(['ws_queue.intent.0001']);
-
-    const strandState = await graph.materializeStrand('ws_queue');
-    const strandReader = createStateReader(strandState);
-    expect(strandReader.getNodeProps('task:queued')).toMatchObject({ status: 'base' });
-
-    const liveState = await graph.materialize();
-    const liveReader = createStateReader(liveState);
-    expect(liveReader.getNodeProps('task:queued')).toMatchObject({ status: 'base' });
-  });
-
-  it('ticks strands deterministically and admits independent intents together', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'task:red', dot: Dot.create('alice', 1) },
-        { type: 'NodeAdd', node: 'task:blue', dot: Dot.create('alice', 2) },
-      ],
-      reads: [],
-      writes: ['task:red', 'task:blue'],
-    });
-
-    await graph.createStrand({ strandId: 'ws_tick', owner: 'alice' });
-
-    await graph.queueStrandIntent('ws_tick', (p) => {
-      p.setProperty('task:red', 'status', 'ready');
-    });
-    await graph.queueStrandIntent('ws_tick', (p) => {
-      p.setProperty('task:blue', 'status', 'review');
-    });
-
-    const result = await graph.tickStrand('ws_tick');
-
-    expect(result.admittedIntentIds).toEqual([
-      'ws_tick.intent.0001',
-      'ws_tick.intent.0002',
-    ]);
-    expect(result.rejected).toEqual([]);
-    expect(result.overlayPatchShas).toHaveLength(2);
-
-    const strandState = await graph.materializeStrand('ws_tick');
-    const strandReader = createStateReader(strandState);
-    expect(strandReader.getNodeProps('task:red')).toMatchObject({ status: 'ready' });
-    expect(strandReader.getNodeProps('task:blue')).toMatchObject({ status: 'review' });
-
-    const descriptor = await graph.getStrand('ws_tick');
-    expect(descriptor?.overlay.patchCount).toBe(2);
-    expect(descriptor?.intentQueue?.intents ?? []).toHaveLength(0);
-    expect(descriptor?.evolution?.tickCount).toBe(1);
-  });
-
-  it('records overlapping queued intents as counterfactuals and leaves sibling lanes untouched', async () => {
-    await simulatePatchCommit(persistence, {
-      graphName,
-      writerId: 'alice',
-      lamport: 1,
-      context: VersionVector.empty(),
-      ops: [
-        { type: 'NodeAdd', node: 'task:conflict', dot: Dot.create('alice', 1) },
-        { type: 'PropSet', node: 'task:conflict', key: 'status', value: 'base' },
-      ],
-      reads: [],
-      writes: ['task:conflict'],
-    });
-
-    await graph.createStrand({ strandId: 'ws_primary', owner: 'alice' });
-    await graph.createStrand({ strandId: 'ws_sibling', owner: 'alice' });
-
-    await graph.queueStrandIntent('ws_primary', (p) => {
-      p.setProperty('task:conflict', 'status', 'approved');
-    });
-    await graph.queueStrandIntent('ws_primary', (p) => {
-      p.setProperty('task:conflict', 'priority', 'urgent');
-    });
-
-    const result = await graph.tickStrand('ws_primary');
-
-    expect(result.admittedIntentIds).toEqual(['ws_primary.intent.0001']);
-    expect(result.rejected).toEqual([
-      {
-        intentId: 'ws_primary.intent.0002',
-        reason: 'footprint_overlap',
-        conflictsWith: ['ws_primary.intent.0001'],
-        reads: ['task:conflict'],
-        writes: ['task:conflict'],
-      },
-    ]);
-
-    const primaryState = await graph.materializeStrand('ws_primary');
-    const primaryReader = createStateReader(primaryState);
-    expect(primaryReader.getNodeProps('task:conflict')).toMatchObject({ status: 'approved' });
-    expect(primaryReader.getNodeProps('task:conflict')).not.toHaveProperty('priority');
-
-    const siblingState = await graph.materializeStrand('ws_sibling');
-    const siblingReader = createStateReader(siblingState);
-    expect(siblingReader.getNodeProps('task:conflict')).toMatchObject({ status: 'base' });
-
-    const liveState = await graph.materialize();
-    const liveReader = createStateReader(liveState);
-    expect(liveReader.getNodeProps('task:conflict')).toMatchObject({ status: 'base' });
-    expect(liveReader.getNodeProps('task:conflict')).not.toHaveProperty('priority');
-  });
 });
